pyproject.toml
[project]
name = "gchat-ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dotenv>=0.9.9",
    "email-validator>=2.2.0",
    "google-adk==0.0.4",
    "google-apps-chat",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "google-cloud",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "google-cloud-firestore",
    "langchain-google-vertexai>=2.0.23",
    "vertexai",
]

[dependency-groups]
dev = [
    "uv-bump>=0.1.2",
]
